commited

+++ MovocashIOS/Features/Home/ACH/FundAccountView.swift
@@ -191,7 +191,7 @@ struct FundAccountView: View {
                         } else {
                             let request = ACHRequest(
                                 source: "manual",
-                                amount: Int(amount) ?? 0,
+                                amount: Double(amount) ?? 0,
                                 achAccountId: account.achAccountId,
                                 userAction: "SUBMITS-ACH-DEPOSIT"
                             )
